# api/openapi.yaml
openapi: 3.0.3
info:
  title: Writing Assistant Engine API
  description: >
    The academic writing-assistant engine's REST/SSE surface (ADR-0017). Typed
    non-streaming routes are codegen'd with ogen; the `/turn` SSE stream is
    hand-framed in the API server (ADR-0031) and is marked x-ogen-raw-response so
    ogen hands the handler the raw http.ResponseWriter.
  version: 0.1.0
servers:
  - url: http://127.0.0.1:9100
paths:
...
  /turn:
    post:
      operationId: startTurn
      summary: Run one agent turn (SSE stream)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Task'
      responses:
        '200':
          description: >
            SSE event stream. Each message is `event: <type>` followed by
            `data: <payload>` where <type> is one of the Event.type enum values
            and <payload> is the matching component schema (TokenEvent,
            MeterEvent, CandidateEvent, DiffEvent, RagEvent, DoneEvent,
            ErrorEvent, BackpressureEvent). One turn per connection: the server
            demultiplexes a turn's events to exactly one client stream, so
            payloads do not repeat the turnId. (ADR-0017 §6, amended.)
          content:
            text/event-stream:
              x-ogen-raw-response: true
              schema:
                $ref: '#/components/schemas/Event'
...
components:
...
  schemas:
...
    Selection:
      type: object
      properties:
        blockId:
          type: string
    TurnOptions:
      type: object
      properties:
        temperature:
          type: number
        model:
          type: string
    Mention:
      type: object
      description: A turn-scoped file attachment (ADR-0036); absolute path resolved by the client.
      required: [path]
      properties:
        path:
          type: string
    Task:
      type: object
      required: [sessionId, modeName, documentId, userInput]
      properties:
        sessionId:
          type: string
        modeName:
          type: string
        documentId:
          type: string
        userInput:
          type: string
        selection:
          $ref: '#/components/schemas/Selection'
        mentions:
          type: array
          items:
            $ref: '#/components/schemas/Mention'
        options:
          $ref: '#/components/schemas/TurnOptions'
...
    Event:
      type: object
      description: >
        Framing marker for the SSE stream: the `event:` line carries `type`;
        the `data:` line carries the payload schema matching that type. The
        payload JSON itself has no `type` field.
      required: [type]
      properties:
        type:
          type: string
          enum: [token, meter, candidate, diff, rag, done, error, backpressure]
...